Move `Board` to its own file
