+++ src/index.js
@@ -2,36 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import MoveLabel from './components/move-label';
 import BoldMoveLabel from './components/bold-move-label';
-import Square from './components/square';
+import Board from './components/board';
 import './index.css';
-
-class Board extends React.Component {
-  renderSquare(i) {
-    return (
-      <Square
-        key={i}
-        value={this.props.squares[i]}
-        onClick={() => this.props.onClick(i)}
-      />
-    );
-  }
-
-  render() {
-    let rows = [];
-    for (let i = 0; i < 3; ++i) {
-      let cells = [];
-      for (let j = 0; j < 3; ++j) {
-        cells.push(this.renderSquare(3 * i + j));
-      }
-      rows.push(
-        <div key={i} className="board-row">{cells}</div>
-      );
-    }
-    return (
-      <div>{rows}</div>
-    );
-  }
-}
 
 class Game extends React.Component {
   constructor() {
